pyproject.toml
[build-system]
requires = [
    "setuptools >= 61.0.0",  # Support for setuptools config in pyproject.toml
]
build-backend = "setuptools.build_meta"

[project]
name = "dpapi-ng"
description = "DPAPI NG decryption for Python"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    { name = "Jordan Borean", email = "jborean93@gmail.com" }
]
keywords = ["dpapi", "dpapi-ng", "laps"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]
dependencies = [
    "dnspython >= 2.0.0",  # Needed for DC lookup support
    "cryptography >= 3.4.4",  # py.typed added, runtime technically need 3.1 but this makes it simpler for testing
    "pyspnego >= 0.9.0",  # Needed for DCE support
]
dynamic = ["version"]

[project.optional-dependencies]
kerberos = [
    "pyspnego[kerberos] >= 0.9.0"
]

[project.urls]
homepage = "https://github.com/jborean93/dpapi-ng"

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "dpapi_ng._version.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sansldap = ["py.typed"]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.mypy]
exclude = "setup.py|build/|docs/|tests/integration/"
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"
show_error_codes = true
show_column_numbers = true
disallow_any_unimported = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = "tests"
junit_family = "xunit2"
norecursedirs = "tests/integration"
